requirements.txt
annotated-types==0.6.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
awscli==1.29.85
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.4
boto==2.49.0
botocore==1.31.85
catboost==1.2.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cloud-init==20.4.1
colorama==0.4.6
comm==0.1.4
configobj==5.0.8
contourpy==1.1.1
crepes==0.6.1
cryptography==41.0.5
cycler==0.11.0
data==0.4
dbus-python==1.2.16
debugpy==1.8.0
decorator==5.1.1
deepspeed==0.12.3
defusedxml==0.7.1
distlib==0.3.6
distro-info==1.0
docutils==0.20.1
einops==0.7.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.1
filelock==3.9.0
fonttools==4.42.1
fqdn==1.5.1
fsspec==2023.10.0
funcsigs==1.0.2
future==0.18.3
graphviz==0.20.1
hjson==3.1.0
httplib2==0.22.0
huggingface-hub==0.17.3
idna==3.4
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflect==7.0.0
ipykernel==6.25.2
ipython==8.15.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
latex==0.7.0
librosa==0.10.1
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==10.1.0
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
ninja==1.11.1.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pipenv==2023.2.4
platformdirs==3.0.0
plotly==5.17.0
progressbar==2.5
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pycparser==2.21
pycurl==7.43.0.6
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.16.1
PyJWT==2.8.0
pynvml==11.5.0
pyparsing==3.1.1
pyrsistent==0.20.0
PySimpleSOAP==1.16.2
python-apt==2.2.1
python-dateutil==2.8.2
python-debian==0.1.49
python-debianbts==4.0.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==5.3.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
regex==2023.10.3
reportbug==7.10.3+deb11u1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==2.0.0
rotary-embedding-torch==0.3.5
rpds-py==0.12.0
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
Send2Trash==1.8.2
shutilwhich==1.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
sympy==1.12
tempdir==0.7.1
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.14.1
tomli==2.0.1
torch==2.1.0
tornado==6.3.3
TorToiSe==2.8.0
tqdm==4.66.1
traitlets==5.10.0
transformers==4.35.0
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
unattended-upgrades==0.1
Unidecode==1.3.7
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.19.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
xgboost==2.0.0
zipp==3.17.0
